Added postgresql commands to constructor, started implementing some functions for sql db retrieval

+++ src/main/java/IEmbedder.java
@@ -1,4 +1,4 @@
 import java.util.List;
 public interface IEmbedder {
-    public List<Float> getEmbedding(String text);
+    public List<Double> getEmbedding(String text);
 }
