+++ src/main/java/ISongSearcher.java
@@ -1,5 +1,5 @@
 import java.util.List;
 
 public interface ISongSearcher {
-    public List<Song> getSimilarSongs(List<Float> embedding); // Embedding of type List<Float>
+    public List<Song> getSimilarSongs(List<Double> embedding); // Embedding of type List<Float>
 }
